+++ ChangeLog.md
@@ -3,6 +3,10 @@ Exchange Webservices change log
 
 ## ?.?.? / ????-??-??
 
+* Added support for timezone context, which is necessary for e.g. Free/Busy.
+  See https://msdn.microsoft.com/en-us/library/office/ff406132(v=exchg.140).aspx
+  (@thekid)
+
 ## 0.1.0 / 2017-01-22
 
 * Hello World! First release - @thekid